+++ fineAnts_app/src/main/java/codesquad/fineants/FineAntsAppApplication.java
@@ -2,8 +2,10 @@ package codesquad.fineants;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing
 public class FineAntsAppApplication {
 
 	public static void main(String[] args) {
