+++ fineAnts_app/src/main/java/codesquad/fineants/domain/member/MemberRepository.java
@@ -5,5 +5,5 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
-	Optional<Member> findMemberByEmailAAndProvider(String email, String provider);
+	Optional<Member> findMemberByEmailAndProvider(String email, String provider);
 }
