+++ fineAnts_app/src/main/java/codesquad/fineants/domain/oauth/client/NaverOauthClient.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.spring.api.oauth.client;
+package codesquad.fineants.domain.oauth.client;
 
 import java.nio.charset.StandardCharsets;
 import java.util.List;
@@ -10,11 +10,11 @@ import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.reactive.function.client.WebClient;
 
+import codesquad.fineants.domain.oauth.properties.OauthProperties;
+import codesquad.fineants.domain.oauth.response.OauthAccessTokenResponse;
+import codesquad.fineants.domain.oauth.response.OauthUserProfileResponse;
 import codesquad.fineants.spring.api.errors.errorcode.OauthErrorCode;
 import codesquad.fineants.spring.api.errors.exception.BadRequestException;
-import codesquad.fineants.spring.api.oauth.properties.OauthProperties;
-import codesquad.fineants.spring.api.oauth.response.OauthAccessTokenResponse;
-import codesquad.fineants.spring.api.oauth.response.OauthUserProfileResponse;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
