+++ fineAnts_app/src/main/java/codesquad/fineants/domain/oauth/client/OauthClient.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.spring.api.oauth.client;
+package codesquad.fineants.domain.oauth.client;
 
 import java.util.Map;
 
@@ -6,8 +6,8 @@ import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.reactive.function.client.WebClient;
 
-import codesquad.fineants.spring.api.oauth.response.OauthAccessTokenResponse;
-import codesquad.fineants.spring.api.oauth.response.OauthUserProfileResponse;
+import codesquad.fineants.domain.oauth.response.OauthAccessTokenResponse;
+import codesquad.fineants.domain.oauth.response.OauthUserProfileResponse;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
