+++ fineAnts_app/src/main/java/codesquad/fineants/domain/oauth/repository/OauthClientRepository.java
@@ -1,0 +1,7 @@
+package codesquad.fineants.domain.oauth.repository;
+
+import codesquad.fineants.domain.oauth.client.OauthClient;
+
+public interface OauthClientRepository {
+	OauthClient findOneBy(String providerName);
+}
