+++ fineAnts_app/src/main/java/codesquad/fineants/domain/oauth/response/OauthAccessTokenResponse.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.spring.api.oauth.response;
+package codesquad.fineants.domain.oauth.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
