+++ fineAnts_app/src/main/java/codesquad/fineants/domain/oauth/response/OauthLoginMemberResponse.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.spring.api.oauth.response;
+package codesquad.fineants.domain.oauth.response;
 
 import codesquad.fineants.domain.member.Member;
 import lombok.AccessLevel;
