+++ fineAnts_app/src/main/java/codesquad/fineants/domain/oauth/support/AuthenticationContext.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.spring.api.oauth.support;
+package codesquad.fineants.domain.oauth.support;
 
 import org.springframework.stereotype.Component;
 import org.springframework.web.context.annotation.RequestScope;
