--- fineAnts_app/src/main/java/codesquad/fineants/spring/api/oauth/repository/OauthClientRepository.java
@@ -1,7 +1,0 @@
-package codesquad.fineants.spring.api.oauth.repository;
-
-import codesquad.fineants.spring.api.oauth.client.OauthClient;
-
-public interface OauthClientRepository {
-	OauthClient findOneBy(String providerName);
-}
