+++ fineAnts_app/src/main/java/codesquad/fineants/spring/config/OauthConfig.java
@@ -4,8 +4,8 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import codesquad.fineants.spring.api.oauth.properties.OauthProperties;
-import codesquad.fineants.spring.api.oauth.repository.InMemoryOauthClientRepository;
+import codesquad.fineants.domain.oauth.properties.OauthProperties;
+import codesquad.fineants.domain.oauth.repository.InMemoryOauthClientRepository;
 import lombok.RequiredArgsConstructor;
 
 @EnableConfigurationProperties(OauthProperties.class)
